Release version release_bl_iot_sdk_1.6.18-31-gb050a94a

+++ components/stage/cli/cli/cli.c
@@ -1010,7 +1010,7 @@ static int cb_idnoe(void *arg, inode_t *node)
         printf("----------------------------------------------------------------------------------\r\n");
     }
     printf("%10d\t\t%30s\t\t\t%s\r\n",
-            INODE_IS_CHAR(node) ? sizeof(struct file_ops) : (INODE_IS_BLOCK(node) ? sizeof(struct file_ops) : (INODE_IS_FS(node) ? sizeof(struct fs_ops) : 0)),
+            (int)(INODE_IS_CHAR(node) ? sizeof(struct file_ops) : (INODE_IS_BLOCK(node) ? sizeof(struct file_ops) : (INODE_IS_FS(node) ? sizeof(struct fs_ops) : 0))),
             node->i_name,
             INODE_IS_CHAR(node) ? "Char Device" : (INODE_IS_BLOCK(node) ? "Block Device" : (INODE_IS_FS(node) ? "File System" : "Unknown"))
     );
@@ -1120,7 +1120,7 @@ static void hexdump_cmd(char *buf, int len, int argc, char **argv)
     printf("Found file %s. XIP Addr %p, len %lu\r\n",
             argv[1],
             filebuf.buf,
-            filebuf.bufsize
+            (unsigned long)filebuf.bufsize
     );
     utils_hexdump(filebuf.buf, filebuf.bufsize);
     aos_close(fd);
